Create folder to each lib and use float
- Arduino doesn't support 8 bytes double.
- Fix WaterTank/WaterSource circular import issue

+++ lib/PressureReader/PressureReader.cpp
@@ -2,11 +2,11 @@
 
 #include "PressureReader.h"
 
-PressureReader::PressureReader(IOInterface* io, double pressureFactor) {
+PressureReader::PressureReader(IOInterface* io, float pressureFactor) {
     this->io = io;
     this->pressureFactor = pressureFactor;
 }
 
-double PressureReader::getValue() {
+float PressureReader::getValue() {
 	return this->io->read() * this->pressureFactor;
 }
